Big misc commit. This commit fixes almost all non #warning type warnings and all uses triggered by the use of the -pedantic compiler flag (requires -Wno-long-long with wx2.5.x though). Also removes around 20 typedefs from Types.h and changes those into the more common types.

+++ trunk/src/ServerListCtrl.cpp
@@ -139,7 +139,7 @@ void CServerListCtrl::RemoveAllServers( int state, bool ask_static )
 	  theApp.serverconnect->IsConnecting();
 
 	while ( pos != -1 ) {
-		if ( GetItemData(pos) == m_connected && connected == true) {
+		if ( (long)GetItemData(pos) == m_connected && connected == true) {
 			wxMessageBox(_("You are connected to a server you are trying to delete. Please disconnect first. The server was NOT deleted."), _("Info"), wxOK);
 			++pos;
 		} else {
